tests: refactored folder structure for better reasoning
